Build Statesman using CircleCI 2.0
This switches to using CircleCI 2.0 to build Statesman.

This involves using a fairly complex Workflows setup with lots of
different jobs for each Ruby version, Rails version and database
adapter. It's a little more complicated to set up than Travis,
since there's no matrix support, but it's faster and more aligned
with our tooling at GoCardless for everything else. We end up
with a mammoth 21 job workflow 🙈🐘

We were previously missing the two database adapters for Rails
5.0 and 5.1, so this also adds those.

It's worth noting that it doesn't seem to be possible to achieve
better reuse (i.e. a more DRY approach) with the different images
in the config (e.g. the standard setup for MySQL, Postgres).
See https://stackoverflow.com/questions/4948933/is-there-a-way-to-alias-anchor-an-array-in-yaml
for details. It doesn't appar that you can have a key that
contains multiple merge keys as an array.